+++ src/low_pass_filter.cpp
@@ -3,13 +3,13 @@
 LowPassFilter::LowPassFilter() {
     m_y = 0;
     m_rc = 0;
-    m_dt = 1 / 44100.0;
+    m_dt = 1 / 44100.0f;
 }
 
 LowPassFilter::~LowPassFilter() {
     /* void */
 }
 
-double LowPassFilter::f(double sample) {
+float LowPassFilter::f(float sample) {
     return fast_f(sample);
 }
